# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 个人技术博客系统 API
  version: v1
  description: |
    个人技术博客系统 OpenAPI 规范。
    基础路径统一为 `/api/v1/**`，可直接导入 Apifox 使用。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/v1/admin/users:
    get:
      tags:
        - Admin Users
      summary: 获取用户分页列表
      description: 权限级别 `ADMIN`。支持按用户名、邮箱、状态和角色组合筛选，同时传入多个条件时按 AND 关系查询。默认按创建时间倒序、用户 ID 倒序返回，不提供自定义排序参数。
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: username
          in: query
          description: 用户名模糊搜索
          required: false
          schema:
            type: string
            maxLength: 20
            example: alice
        - name: email
          in: query
          description: 邮箱模糊搜索
          required: false
          schema:
            type: string
            maxLength: 255
            example: example.com
        - name: role
          in: query
          description: 角色筛选
          required: false
          schema:
            type: string
            enum:
              - ADMIN
              - USER
            example: USER
        - name: status
          in: query
          description: 状态筛选
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - DISABLED
            example: ACTIVE
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserPageResponse'
              example:
                code: 0
                message: 成功
                data:
                  records:
                    - id: 10002
                      username: alice_dev
                      nickname: Alice
                      email: alice@example.com
                      role: USER
                      status: ACTIVE
                      lastLoginAt: '2026-04-22T22:20:10+08:00'
                      createdAt: '2026-04-22T22:20:00+08:00'
                  pageNum: 1
                  pageSize: 10
                  total: 1
                  totalPages: 1
                  hasNext: false
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
  /api/v1/admin/users/{userId}/status:
    patch:
      tags:
        - Admin Users
      summary: 修改用户状态
      description: 权限级别 `ADMIN`。支持启用或禁用用户；管理员不能修改当前登录用户自身状态。状态修改为 DISABLED 时会撤销该用户全部活跃 Refresh Token，并递增 tokenVersion 使旧 Access Token 立即失效；修改为 ACTIVE 时不撤销 Refresh Token，也不再次递增 tokenVersion。
      operationId: updateAdminUserStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
            example:
              status: DISABLED
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserStatusResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 10002
                  status: DISABLED
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: 无权限访问或不允许修改当前登录用户自身状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  summary: 无权限访问
                  value:
                    code: 101003
                    message: 无权限访问
                    data: null
                selfStatusChangeNotAllowed:
                  summary: 不允许修改当前登录用户状态
                  value:
                    code: 102008
                    message: 当前用户不允许修改自己的状态
                    data: null
  /api/v1/admin/users/{userId}/role:
    patch:
      tags:
        - Admin Users
      summary: 修改用户角色
      description: 权限级别 `ADMIN`。支持将用户修改为管理员或普通用户；管理员不能修改当前登录用户自身角色。角色发生变化后不直接撤销 Refresh Token，但会递增 tokenVersion 使旧 Access Token 立即失效；新的角色在刷新登录态或重新登录后生效。
      operationId: updateAdminUserRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
            example:
              role: ADMIN
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserRoleResponse'
              example:
                code: 0
                message: 成功
                data:
                  id: 10002
                  role: ADMIN
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: 无权限访问或不允许修改当前登录用户自身角色
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  summary: 无权限访问
                  value:
                    code: 101003
                    message: 无权限访问
                    data: null
                selfRoleChangeNotAllowed:
                  summary: 不允许修改当前登录用户角色
                  value:
                    code: 102007
                    message: 当前用户不允许修改自己的角色
                    data: null
...
  /api/v1/admin/messages:
    get:
      tags:
        - Admin Messages
      summary: 获取后台留言分页列表
      operationId: listAdminMessages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: messageId
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: guestNickname
          in: query
          schema:
            type: string
            maxLength: 20
        - name: guestEmail
          in: query
          schema:
            type: string
            maxLength: 255
        - name: content
          in: query
          schema:
            type: string
            maxLength: 1000
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/MessageStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/MessageType'
        - name: createdAtFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: createdAtTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMessagePageResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
  /api/v1/admin/messages/{messageId}/moderation:
    patch:
      tags:
        - Admin Messages
      summary: 审核、隐藏或删除留言
      operationId: moderateMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageModerationRequest'
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageMutationResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '400':
          $ref: '#/components/responses/MessageBadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/MessageNotFoundResponse'
        '409':
          $ref: '#/components/responses/MessageConflictResponse'
  /api/v1/admin/messages/{messageId}/replies:
    post:
      tags:
        - Admin Messages
      summary: 管理员回复留言
      operationId: replyMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageReplyRequest'
      responses:
        '200':
          description: 回复成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageMutationResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '400':
          $ref: '#/components/responses/MessageBadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/MessageNotFoundResponse'
        '409':
          $ref: '#/components/responses/MessageConflictResponse'
  /api/v1/admin/messages/batch-approval:
    patch:
      tags:
        - Admin Messages
      summary: 批量通过待审核顶层留言
      operationId: approveMessages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageBatchApprovalRequest'
      responses:
        '200':
          $ref: '#/components/responses/NoDataResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '400':
          $ref: '#/components/responses/MessageBadRequestResponse'
        '409':
          $ref: '#/components/responses/MessageConflictResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNum:
      name: pageNum
      in: query
      description: 当前页码，从 1 开始，默认 1
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSize:
      name: pageSize
      in: query
      description: 每页条数，默认 10
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 20
        default: 10
        example: 10
...
    UserId:
      name: userId
      in: path
      description: 用户 ID
      required: true
      schema:
        type: integer
        format: int64
        example: 10002
...
    MessageId:
      name: messageId
      in: path
      description: 留言 ID
      required: true
      schema:
        type: integer
        format: int64
        example: 90001
...
  schemas:
    BaseResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 业务状态码，`0` 表示成功
          example: 0
        message:
          type: string
          description: 业务描述信息
          example: 成功
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              nullable: true
              description: 错误场景通常为空
              example: null
...
    PagedMeta:
      type: object
      required:
        - pageNum
        - pageSize
        - total
        - totalPages
        - hasNext
      properties:
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 35
        totalPages:
          type: integer
          example: 4
        hasNext:
          type: boolean
          example: true
...
    AdminUserListItem:
      type: object
      required:
        - id
        - username
        - nickname
        - email
        - role
        - status
        - lastLoginAt
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 10002
        username:
          type: string
          example: alice_dev
        nickname:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
        role:
          type: string
          enum:
            - ADMIN
            - USER
          example: USER
        status:
          type: string
          enum:
            - ACTIVE
            - DISABLED
          example: ACTIVE
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-04-22T22:20:10+08:00'
        createdAt:
          type: string
          format: date-time
          example: '2026-04-22T22:20:00+08:00'
    AdminUserPageData:
      allOf:
        - $ref: '#/components/schemas/PagedMeta'
        - type: object
          required:
            - records
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/AdminUserListItem'
    AdminUserPageResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminUserPageData'
    UpdateUserStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ACTIVE
            - DISABLED
          example: DISABLED
    UpdateUserStatusData:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: integer
          format: int64
          example: 10002
        status:
          type: string
          enum:
            - ACTIVE
            - DISABLED
          example: DISABLED
    UpdateUserStatusResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UpdateUserStatusData'
    UpdateUserRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - ADMIN
            - USER
          example: ADMIN
    UpdateUserRoleData:
      type: object
      required:
        - id
        - role
      properties:
        id:
          type: integer
          format: int64
          example: 10002
        role:
          type: string
          enum:
            - ADMIN
            - USER
          example: ADMIN
    UpdateUserRoleResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UpdateUserRoleData'
...
    MessageStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, HIDDEN, DELETED]
    MessageType:
      type: string
      enum: [TOP_LEVEL, REPLY]
    MessageAuthor:
      type: object
      required: [id, username, nickname, avatarUrl]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          nullable: true
    MessageReply:
      type: object
      required: [id, parentId, content, status, author, createdAt]
      properties:
        id:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
        content:
          type: string
        status:
          $ref: '#/components/schemas/MessageStatus'
        author:
          allOf:
            - $ref: '#/components/schemas/MessageAuthor'
          nullable: true
        createdAt:
          type: string
          format: date-time
    MessageItem:
      type: object
      required: [id, nickname, content, status, moderationReason, author, isMine, replies, createdAt]
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        content:
          type: string
        status:
          $ref: '#/components/schemas/MessageStatus'
        moderationReason:
          type: string
          nullable: true
        author:
          allOf:
            - $ref: '#/components/schemas/MessageAuthor'
          nullable: true
        isMine:
          type: boolean
        replies:
          type: array
          items:
            $ref: '#/components/schemas/MessageReply'
        createdAt:
          type: string
          format: date-time
...
    MessageMutationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MessageMutationData'
    MessageMutationData:
      allOf:
        - $ref: '#/components/schemas/MessageItem'
        - type: object
          required: [parentId]
          properties:
            parentId:
              type: integer
              format: int64
              nullable: true
...
    MessageModerationRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [APPROVE, REJECT, HIDE, DELETE]
        reason:
          type: string
          maxLength: 255
    CreateMessageReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
    MessageBatchApprovalRequest:
      type: object
      required: [messageIds]
      properties:
        messageIds:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            type: integer
            format: int64
            minimum: 1
    AdminMessageItem:
      type: object
      required:
        - id
        - userId
        - parentId
        - nickname
        - email
        - content
        - status
        - type
        - notifyOnReply
        - moderationReason
        - reviewedBy
        - reviewedAt
        - deletedBy
        - deletedAt
        - author
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
          nullable: true
        parentId:
          type: integer
          format: int64
          nullable: true
        nickname:
          type: string
        email:
          type: string
          description: 仅后台返回的私有联系邮箱
        content:
          type: string
        status:
          $ref: '#/components/schemas/MessageStatus'
        type:
          $ref: '#/components/schemas/MessageType'
        notifyOnReply:
          type: boolean
        moderationReason:
          type: string
          nullable: true
        reviewedBy:
          type: integer
          format: int64
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        deletedBy:
          type: integer
          format: int64
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        author:
          allOf:
            - $ref: '#/components/schemas/MessageAuthor'
          nullable: true
        createdAt:
          type: string
          format: date-time
    AdminMessagePageData:
      allOf:
        - $ref: '#/components/schemas/PagedMeta'
        - type: object
          required: [records]
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/AdminMessageItem'
    AdminMessagePageResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminMessagePageData'
  responses:
    BadRequestResponse:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 199001
            message: 请求参数不合法
            data: null
    UnauthorizedResponse:
      description: 未登录或 Access Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 101001
            message: 未登录或 Access Token 无效
            data: null
    ForbiddenResponse:
      description: 无权限访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 101003
            message: 无权限访问
            data: null
...
    MessageNotFoundResponse:
      description: 留言不存在、已删除或当前用户不可见
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 108001
            message: 留言不存在
            data: null
    MessageBadRequestResponse:
      description: 留言请求参数或业务参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidParameter:
              value:
                code: 199001
                message: 请求参数不合法
                data: null
            invalidNickname:
              value:
                code: 108008
                message: 留言昵称不合法
                data: null
            emailRequired:
              value:
                code: 108009
                message: 勾选回复通知时必须填写邮箱
                data: null
            invalidUnsubscribeToken:
              value:
                code: 108010
                message: 留言退订链接无效或已失效
                data: null
            batchTooLarge:
              value:
                code: 108012
                message: 一次最多通过 100 条留言
                data: null
    MessageConflictResponse:
      description: 留言状态或回复目标冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            parentUnavailable:
              value:
                code: 108003
                message: 留言回复目标不可用
                data: null
            statusInvalid:
              value:
                code: 108006
                message: 留言状态流转不合法
                data: null
            batchInvalid:
              value:
                code: 108011
                message: 批量通过的留言必须是待审核顶层留言
                data: null
...